+++ src/email/email.test.ts
@@ -3,10 +3,10 @@
 import { createValidator } from '../createValidator/createValidator'
 import { stringYobta } from '../stringYobta'
 import { YobtaError } from '../YobtaError'
-import { emailMessage, emailYobta } from './'
+import { email, emailMessage } from './email'
 
 const customMessage = 'yobta!'
-const validate = createValidator(stringYobta(), emailYobta(customMessage))
+const validate = createValidator(stringYobta(), email(customMessage))
 
 it(`accepts valid emails`, async () => {
   const result = validate('user-@example.org')
@@ -30,7 +30,7 @@ it(`rejects invalid email`, () => {
 })
 
 it('has default error message', () => {
-  const rule = emailYobta()
+  const rule = email()
   const validateDefault = createValidator(rule)
   const attempt = (): any => validateDefault('yobta')
   expect(attempt).toThrow(emailMessage)