+++ src/email/email.ts
@@ -1,17 +1,17 @@
-import { reEmailYobta } from '../_patterns/reEmailYobta.js'
 import type { YobtaSyncRule } from '../ruleYobta/index.js'
 import { ruleYobta } from '../ruleYobta/index.js'
+import { reEmail } from './reEmail.js'
 
 export const emailMessage = 'It should be an email'
 
 export type YobtaEmail = { __email__: void } & string
 
-export const emailYobta = <I extends string>(
+export const email = <I extends string>(
   message = emailMessage,
 ): YobtaSyncRule<I, YobtaEmail> =>
   ruleYobta((value: I) => {
     const trimmed = value.trim()
-    if (reEmailYobta.test(trimmed)) {
+    if (reEmail.test(trimmed)) {
       return trimmed as YobtaEmail
     }
     throw new Error(message)