+++ src/email/index.ts
@@ -1,0 +1,2 @@
+export * from './email.js'
+export * from './reEmail.js'